virtus-assistant/requirements.txt
# VIRTUS AI Asistan - Tam Bağımlılıklar
# Tüm özellikler için gerekli paketler

# ============================================
# CORE - Temel Paketler
# ============================================
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0         # Web scraping (araştırma için)
lxml>=4.9.0                    # HTML parsing

# ============================================
# SPEECH RECOGNITION - Ses Tanıma
# ============================================
SpeechRecognition>=3.10.0

# PyAudio - MİKROFON İÇİN GEREKLİ!
# Windows için özel kurulum:
#   pip install pipwin
#   python -m pipwin install pyaudio
# Linux için:
#   sudo apt-get install portaudio19-dev python3-pyaudio
#   pip install pyaudio
# Mac için:
#   brew install portaudio
#   pip install pyaudio

# Ses dosyası işleme
pydub>=0.25.1                  # MP3 desteği için (PyAudio ile)
sounddevice>=0.4.6             # Alternatif mikrofon
numpy>=1.24.0                  # sounddevice için

# ============================================
# TEXT-TO-SPEECH - Ses Çıkışı
# ============================================

# Online TTS (En İyi Kalite)
gTTS>=2.5.0                    # Google TTS - Doğal erkek sesi
azure-cognitiveservices-speech>=1.31.0  # Azure TTS (opsiyonel, premium)

# Offline TTS (Fallback)
pyttsx3>=2.90

# Ses Çalma - playsound yerine daha iyi alternatifler
pygame>=2.5.0                  # Cross-platform (öncelikli)
pydub>=0.25.1                  # MP3 desteği için

# ============================================
# AI - Google Gemini
# ============================================
google-generativeai>=0.3.0
# veya yeni API:
google-genai>=0.2.0

# ============================================
# SYSTEM CONTROL - Sistem Kontrolü
# ============================================
pyautogui>=0.9.54
psutil>=5.9.0

# Windows Specific
pywin32>=305; sys_platform == 'win32'
comtypes>=1.2.0; sys_platform == 'win32'
pycaw>=20181226; sys_platform == 'win32'  # Ses kontrolü
wmi>=1.5.1; sys_platform == 'win32'       # Sistem bilgisi

# ============================================
# WAKE WORD DETECTION (Opsiyonel)
# ============================================
# pvporcupine>=3.0.0  # Picovoice Porcupine (API key gerekli)
# Ücretsiz alternatif: snowboy (deprecated ama çalışıyor)

# ============================================
# MOBILE & CROSS-PLATFORM (Opsiyonel)
# ============================================
# kivy==2.2.1          # Mobil app için
# plyer==2.1.0         # Android özellikleri
# buildozer==1.5.0     # Android build

# ============================================
# UTILITIES
# ============================================
colorama>=0.4.6      # Renkli terminal